Struct mailpot::db::Connection

pub struct Connection {
    pub connection: DbConnection,
    conf: Configuration,
}

A connection to a mailpot database.

Fields
connection: DbConnection

The rusqlite connection handle.

conf: Configuration
Implementations

impl Connection

pub fn fetch_templates(&self) -> Result<Vec<DbVal<Template>>>

Fetch all.

pub fn fetch_template( &self, template: &str, list_pk: Option<i64> ) -> Result<Option<DbVal<Template>>>

Fetch a named template.

pub fn add_template(&self, template: Template) -> Result<DbVal<Template>>

Insert a named template.

pub fn remove_template( &self, template: &str, list_pk: Option<i64> ) -> Result<Template>

Remove a named template.

impl Connection

pub fn insert_to_queue(&self, entry: QueueEntry) -> Result<DbVal<QueueEntry>>

Insert a received email into a queue.

pub fn queue(&self, queue: Queue) -> Result<Vec<DbVal<QueueEntry>>>

Fetch all queue entries.

pub fn delete_from_queue( &mut self, queue: Queue, index: Vec<i64> ) -> Result<Vec<QueueEntry>>

Delete queue entries returning the deleted values.

impl Connection

pub fn insert_post( &self, list_pk: i64, message: &[u8], env: &Envelope ) -> Result<i64>

Insert a mailing list post into the database.

pub fn post(&mut self, env: &Envelope, raw: &[u8], _dry_run: bool) -> Result<()>

Process a new mailing list post.

fn inner_post( &mut self, env: &Envelope, raw: &[u8], _dry_run: bool ) -> Result<()>

pub fn request( &mut self, list: &DbVal<MailingList>, request: ListRequest, env: &Envelope, raw: &[u8] ) -> Result<()>

Process a new mailing list request.

pub fn months(&self, list_pk: i64) -> Result<Vec<String>>

Fetch all year and month values for which at least one post exists in yyyy-mm format.

pub fn list_post_by_message_id( &self, list_pk: i64, message_id: &str ) -> Result<Option<DbVal<Post>>>

Find a post by its Message-ID email header.

pub fn send_reply_with_list_template<'ctx, F: Fn() -> Template>( &self, render_context: TemplateRenderContext<'ctx, F>, recipients: impl Iterator<Item = Cow<'ctx, Address>> ) -> Result<()>

Helper function to send a template reply.

impl Connection

pub fn list_subscriptions( &self, pk: i64 ) -> Result<Vec<DbVal<ListSubscription>>>

Fetch all subscriptions of a mailing list.

pub fn list_subscription( &self, list_pk: i64, pk: i64 ) -> Result<DbVal<ListSubscription>>

Fetch mailing list subscription.

pub fn list_subscription_by_address( &self, list_pk: i64, address: &str ) -> Result<DbVal<ListSubscription>>

Fetch mailing list subscription by their address.

pub fn add_subscription( &self, list_pk: i64, new_val: ListSubscription ) -> Result<DbVal<ListSubscription>>

Add subscription to mailing list.

pub fn add_candidate_subscription( &self, list_pk: i64, new_val: ListSubscription ) -> Result<DbVal<ListCandidateSubscription>>

Create subscription candidate.

pub fn candidate_subscription( &self, pk: i64 ) -> Result<DbVal<ListCandidateSubscription>>

Fetch subscription candidate by primary key.

pub fn accept_candidate_subscription( &mut self, pk: i64 ) -> Result<DbVal<ListSubscription>>

Accept subscription candidate.

pub fn remove_subscription(&self, list_pk: i64, address: &str) -> Result<()>

Remove a subscription by their address.

pub fn update_subscription( &mut self, change_set: ListSubscriptionChangeset ) -> Result<()>

Update a mailing list subscription.

pub fn account(&self, pk: i64) -> Result<Option<DbVal<Account>>>

Fetch account by pk.

pub fn account_by_address( &self, address: &str ) -> Result<Option<DbVal<Account>>>

Fetch account by address.

pub fn account_subscriptions( &self, pk: i64 ) -> Result<Vec<DbVal<ListSubscription>>>

Fetch all subscriptions of an account by primary key.

pub fn accounts(&self) -> Result<Vec<DbVal<Account>>>

Fetch all accounts.

pub fn add_account(&self, new_val: Account) -> Result<DbVal<Account>>

Add account.

pub fn remove_account(&self, address: &str) -> Result<()>

Remove an account by their address.

pub fn update_account(&mut self, change_set: AccountChangeset) -> Result<()>

Update an account.

impl Connection

pub fn list_post_policy(&self, pk: i64) -> Result<Option<DbVal<PostPolicy>>>

Fetch the post policy of a mailing list.

pub fn remove_list_post_policy( &self, list_pk: i64, policy_pk: i64 ) -> Result<()>

Remove an existing list policy.

let db = Connection::open_or_create_db(config).unwrap().trusted();
let list = db
    .create_list(MailingList {
        pk: 0,
        name: "foobar chat".into(),
        id: "foo-chat".into(),
        address: "foo-chat@example.com".into(),
        description: None,
        archive_url: None,
    })
    .unwrap();

let pol = db
    .set_list_post_policy(PostPolicy {
        pk: -1,
        list: list.pk(),
        announce_only: false,
        subscription_only: true,
        approval_needed: false,
        open: false,
        custom: false,
    })
    .unwrap();
db.remove_list_post_policy(list.pk(), pol.pk()).unwrap();

pub fn remove_list_post_policy_panic()

let db = Connection::open_or_create_db(config).unwrap().trusted();
db.remove_list_post_policy(1, 1).unwrap();

pub fn set_list_post_policy( &self, policy: PostPolicy ) -> Result<DbVal<PostPolicy>>

Set the unique post policy for a list.

impl Connection

pub fn list_subscription_policy( &self, pk: i64 ) -> Result<Option<DbVal<SubscriptionPolicy>>>

Fetch the subscription policy of a mailing list.

pub fn remove_list_subscription_policy( &self, list_pk: i64, policy_pk: i64 ) -> Result<()>

Remove an existing subscription policy.

let db = Connection::open_or_create_db(config).unwrap().trusted();
let list = db
    .create_list(MailingList {
        pk: 0,
        name: "foobar chat".into(),
        id: "foo-chat".into(),
        address: "foo-chat@example.com".into(),
        description: None,
        archive_url: None,
    })
    .unwrap();
let pol = db
    .set_list_subscription_policy(SubscriptionPolicy {
        pk: -1,
        list: list.pk(),
        send_confirmation: false,
        open: true,
        manual: false,
        request: false,
        custom: false,
    })
    .unwrap();
db.remove_list_subscription_policy(list.pk(), pol.pk())
    .unwrap();

pub fn remove_list_subscription_policy_panic()

let db = Connection::open_or_create_db(config).unwrap().trusted();
db.remove_list_post_policy(1, 1).unwrap();

pub fn set_list_subscription_policy( &self, policy: SubscriptionPolicy ) -> Result<DbVal<SubscriptionPolicy>>

Set the unique post policy for a list.

impl Connection

pub const SCHEMA: &str = _

The database schema.

PRAGMA foreign_keys = true;
PRAGMA encoding = 'UTF-8';

CREATE TABLE IF NOT EXISTS list (
  pk                    INTEGER PRIMARY KEY NOT NULL,
  name                  TEXT NOT NULL,
  id                    TEXT NOT NULL UNIQUE,
  address               TEXT NOT NULL UNIQUE,
  owner_local_part      TEXT,
  request_local_part    TEXT,
  archive_url           TEXT,
  description           TEXT,
  created               INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified         INTEGER NOT NULL DEFAULT (unixepoch()),
  verify                BOOLEAN CHECK (verify IN (0, 1)) NOT NULL DEFAULT 1,
  hidden                BOOLEAN CHECK (hidden IN (0, 1)) NOT NULL DEFAULT 0,
  enabled               BOOLEAN CHECK (enabled IN (0, 1)) NOT NULL DEFAULT 1
);

CREATE TABLE IF NOT EXISTS owner (
  pk               INTEGER PRIMARY KEY NOT NULL,
  list             INTEGER NOT NULL,
  address          TEXT NOT NULL,
  name             TEXT,
  created          INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified    INTEGER NOT NULL DEFAULT (unixepoch()),
  FOREIGN KEY (list) REFERENCES list(pk) ON DELETE CASCADE
);

CREATE TABLE IF NOT EXISTS post_policy (
  pk                   INTEGER PRIMARY KEY NOT NULL,
  list                 INTEGER NOT NULL UNIQUE,
  announce_only        BOOLEAN CHECK (announce_only IN (0, 1)) NOT NULL
                       DEFAULT 0,
  subscription_only    BOOLEAN CHECK (subscription_only IN (0, 1)) NOT NULL
                       DEFAULT 0,
  approval_needed      BOOLEAN CHECK (approval_needed IN (0, 1)) NOT NULL
                       DEFAULT 0,
  open                 BOOLEAN CHECK (open IN (0, 1)) NOT NULL DEFAULT 0,
  custom               BOOLEAN CHECK (custom IN (0, 1)) NOT NULL DEFAULT 0,
  created              INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified        INTEGER NOT NULL DEFAULT (unixepoch())
  CHECK((
    (custom) OR ((
    (open) OR ((
    (approval_needed) OR ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  )
  AND NOT
  (
    (approval_needed) AND ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  ))
  )
  AND NOT
  (
    (open) AND ((
    (approval_needed) OR ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  )
  AND NOT
  (
    (approval_needed) AND ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  ))
  ))
  )
  AND NOT
  (
    (custom) AND ((
    (open) OR ((
    (approval_needed) OR ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  )
  AND NOT
  (
    (approval_needed) AND ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  ))
  )
  AND NOT
  (
    (open) AND ((
    (approval_needed) OR ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  )
  AND NOT
  (
    (approval_needed) AND ((
    (announce_only) OR (subscription_only)
  )
  AND NOT
  (
    (announce_only) AND (subscription_only)
  ))
  ))
  ))
  )),
  FOREIGN KEY (list) REFERENCES list(pk) ON DELETE CASCADE
);

CREATE TABLE IF NOT EXISTS subscription_policy (
  pk                   INTEGER PRIMARY KEY NOT NULL,
  list                 INTEGER NOT NULL UNIQUE,
  send_confirmation    BOOLEAN CHECK (send_confirmation IN (0, 1)) NOT NULL
                       DEFAULT 1,
  open                 BOOLEAN CHECK (open IN (0, 1)) NOT NULL DEFAULT 0,
  manual               BOOLEAN CHECK (manual IN (0, 1)) NOT NULL DEFAULT 0,
  request              BOOLEAN CHECK (request IN (0, 1)) NOT NULL DEFAULT 0,
  custom               BOOLEAN CHECK (custom IN (0, 1)) NOT NULL DEFAULT 0,
  created              INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified        INTEGER NOT NULL DEFAULT (unixepoch()),
  CHECK((
    (open) OR ((
    (manual) OR ((
    (request) OR (custom)
  )
  AND NOT
  (
    (request) AND (custom)
  ))
  )
  AND NOT
  (
    (manual) AND ((
    (request) OR (custom)
  )
  AND NOT
  (
    (request) AND (custom)
  ))
  ))
  )
  AND NOT
  (
    (open) AND ((
    (manual) OR ((
    (request) OR (custom)
  )
  AND NOT
  (
    (request) AND (custom)
  ))
  )
  AND NOT
  (
    (manual) AND ((
    (request) OR (custom)
  )
  AND NOT
  (
    (request) AND (custom)
  ))
  ))
  )),
  FOREIGN KEY (list) REFERENCES list(pk) ON DELETE CASCADE
);

CREATE TABLE IF NOT EXISTS subscription (
  pk                      INTEGER PRIMARY KEY NOT NULL,
  list                    INTEGER NOT NULL,
  address                 TEXT NOT NULL,
  name                    TEXT,
  account                 INTEGER,
  enabled                 BOOLEAN CHECK (enabled IN (0, 1)) NOT NULL
                          DEFAULT 1,
  verified                BOOLEAN CHECK (verified IN (0, 1)) NOT NULL
                          DEFAULT 1,
  digest                  BOOLEAN CHECK (digest IN (0, 1)) NOT NULL
                          DEFAULT 0,
  hide_address            BOOLEAN CHECK (hide_address IN (0, 1)) NOT NULL
                          DEFAULT 0,
  receive_duplicates      BOOLEAN CHECK (receive_duplicates IN (0, 1)) NOT NULL
                          DEFAULT 1,
  receive_own_posts       BOOLEAN CHECK (receive_own_posts IN (0, 1)) NOT NULL
                          DEFAULT 0,
  receive_confirmation    BOOLEAN CHECK (receive_confirmation IN (0, 1)) NOT NULL
                          DEFAULT 1,
  last_digest             INTEGER NOT NULL DEFAULT (unixepoch()),
  created                 INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified           INTEGER NOT NULL DEFAULT (unixepoch()),
  FOREIGN KEY (list) REFERENCES list(pk) ON DELETE CASCADE,
  FOREIGN KEY (account) REFERENCES account(pk) ON DELETE SET NULL,
  UNIQUE (list, address) ON CONFLICT ROLLBACK
);

CREATE TABLE IF NOT EXISTS account (
  pk               INTEGER PRIMARY KEY NOT NULL,
  name             TEXT,
  address          TEXT NOT NULL UNIQUE,
  public_key       TEXT,
  password         TEXT NOT NULL,
  enabled          BOOLEAN CHECK (enabled IN (0, 1)) NOT NULL DEFAULT 1,
  created          INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified    INTEGER NOT NULL DEFAULT (unixepoch())
);

CREATE TABLE IF NOT EXISTS candidate_subscription (
  pk               INTEGER PRIMARY KEY NOT NULL,
  list             INTEGER NOT NULL,
  address          TEXT NOT NULL,
  name             TEXT,
  accepted         INTEGER UNIQUE,
  created          INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified    INTEGER NOT NULL DEFAULT (unixepoch()),
  FOREIGN KEY (list) REFERENCES list(pk) ON DELETE CASCADE,
  FOREIGN KEY (accepted) REFERENCES subscription(pk) ON DELETE CASCADE,
  UNIQUE (list, address) ON CONFLICT ROLLBACK
);

CREATE TABLE IF NOT EXISTS post (
  pk               INTEGER PRIMARY KEY NOT NULL,
  list             INTEGER NOT NULL,
  envelope_from    TEXT,
  address          TEXT NOT NULL,
  message_id       TEXT NOT NULL,
  message          BLOB NOT NULL,
  headers_json     TEXT,
  timestamp        INTEGER NOT NULL DEFAULT (unixepoch()),
  datetime         TEXT NOT NULL DEFAULT (datetime()),
  created          INTEGER NOT NULL DEFAULT (unixepoch())
);

CREATE TABLE IF NOT EXISTS templates (
  pk               INTEGER PRIMARY KEY NOT NULL,
  name             TEXT NOT NULL,
  list             INTEGER,
  subject          TEXT,
  headers_json     TEXT,
  body             TEXT NOT NULL,
  created          INTEGER NOT NULL DEFAULT (unixepoch()),
  last_modified    INTEGER NOT NULL DEFAULT (unixepoch()),
  FOREIGN KEY (list) REFERENCES list(pk) ON DELETE CASCADE,
  UNIQUE (list, name) ON CONFLICT ROLLBACK
);

-- # Queues
--
-- ## The "maildrop" queue
--
-- Messages that have been submitted but not yet processed, await processing
-- in the "maildrop" queue. Messages can be added to the "maildrop" queue
-- even when mailpot is not running.
--
-- ## The "deferred" queue
--
-- When all the deliverable recipients for a message are delivered, and for
-- some recipients delivery failed for a transient reason (it might succeed
-- later), the message is placed in the "deferred" queue.
--
-- ## The "hold" queue
--
-- List administrators may introduce rules for emails to be placed
-- indefinitely in the "hold" queue. Messages placed in the "hold" queue stay
-- there until the administrator intervenes. No periodic delivery attempts
-- are made for messages in the "hold" queue.

-- ## The "out" queue
--
-- Emails that must be sent as soon as possible.
CREATE TABLE IF NOT EXISTS queue (
  pk              INTEGER PRIMARY KEY NOT NULL,
  which           TEXT
                  CHECK (
                    which IN
                    ('maildrop',
                     'hold',
                     'deferred',
                     'corrupt',
                     'error',
                     'out')
                  ) NOT NULL,
  list            INTEGER,
  comment         TEXT,
  to_addresses    TEXT NOT NULL,
  from_address    TEXT NOT NULL,
  subject         TEXT NOT NULL,
  message_id      TEXT NOT NULL,
  message         BLOB NOT NULL,
  timestamp       INTEGER NOT NULL DEFAULT (unixepoch()),
  datetime        TEXT NOT NULL DEFAULT (datetime()),
  FOREIGN KEY (list) REFERENCES list(pk) ON DELETE CASCADE,
  UNIQUE (to_addresses, message_id) ON CONFLICT ROLLBACK
);

CREATE TABLE IF NOT EXISTS bounce (
  pk              INTEGER PRIMARY KEY NOT NULL,
  subscription    INTEGER NOT NULL UNIQUE,
  count           INTEGER NOT NULL DEFAULT 0,
  last_bounce     TEXT NOT NULL DEFAULT (datetime()),
  FOREIGN KEY (subscription) REFERENCES subscription(pk) ON DELETE CASCADE
);

CREATE INDEX IF NOT EXISTS post_listpk_idx ON post(list);
CREATE INDEX IF NOT EXISTS post_msgid_idx ON post(message_id);
CREATE INDEX IF NOT EXISTS list_idx ON list(id);
CREATE INDEX IF NOT EXISTS subscription_idx ON subscription(address);

-- [tag:accept_candidate]: Update candidacy with 'subscription' foreign key on
-- 'subscription' insert.
CREATE TRIGGER IF NOT EXISTS accept_candidate AFTER INSERT ON subscription
FOR EACH ROW
BEGIN
  UPDATE candidate_subscription SET accepted = NEW.pk, last_modified = unixepoch()
  WHERE candidate_subscription.list = NEW.list AND candidate_subscription.address = NEW.address;
END;

-- [tag:verify_subscription_email]: If list settings require e-mail to be
-- verified, update new subscription's 'verify' column value.
CREATE TRIGGER IF NOT EXISTS verify_subscription_email AFTER INSERT ON subscription
FOR EACH ROW
BEGIN
  UPDATE subscription
  SET verified = 0, last_modified = unixepoch()
  WHERE
  subscription.pk = NEW.pk
  AND
  EXISTS
  (SELECT 1 FROM list WHERE pk = NEW.list AND verify = 1);
END;

-- [tag:add_account]: Update list subscription entries with 'account' foreign
-- key, if addresses match.
CREATE TRIGGER IF NOT EXISTS add_account AFTER INSERT ON account
FOR EACH ROW
BEGIN
  UPDATE subscription SET account = NEW.pk, last_modified = unixepoch()
  WHERE subscription.address = NEW.address;
END;

-- [tag:add_account_to_subscription]: When adding a new 'subscription', auto
-- set 'account' value if there already exists an 'account' entry with the
-- same address.
CREATE TRIGGER IF NOT EXISTS add_account_to_subscription
AFTER INSERT ON subscription
FOR EACH ROW
WHEN
  NEW.account IS NULL
  AND EXISTS (SELECT 1 FROM account WHERE address = NEW.address)
BEGIN
  UPDATE subscription
     SET account = (SELECT pk FROM account WHERE address = NEW.address),
         last_modified = unixepoch()
    WHERE subscription.pk = NEW.pk;
END;

-- [tag:last_modified_list]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_list
AFTER UPDATE ON list
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE list SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

-- [tag:last_modified_owner]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_owner
AFTER UPDATE ON owner
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE owner SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

-- [tag:last_modified_post_policy]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_post_policy
AFTER UPDATE ON post_policy
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE post_policy SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

-- [tag:last_modified_subscription_policy]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_subscription_policy
AFTER UPDATE ON subscription_policy
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE subscription_policy SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

-- [tag:last_modified_subscription]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_subscription
AFTER UPDATE ON subscription
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE subscription SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

-- [tag:last_modified_account]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_account
AFTER UPDATE ON account
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE account SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

-- [tag:last_modified_candidate_subscription]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_candidate_subscription
AFTER UPDATE ON candidate_subscription
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE candidate_subscription SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

-- [tag:last_modified_templates]: update last_modified on every change.
CREATE TRIGGER
IF NOT EXISTS last_modified_templates
AFTER UPDATE ON templates
FOR EACH ROW
WHEN NEW.last_modified != OLD.last_modified
BEGIN
  UPDATE templates SET last_modified = unixepoch()
  WHERE pk = NEW.pk;
END;

pub fn open_db(conf: Configuration) -> Result<Self>

Creates a new database connection.

Connection supports a limited subset of operations by default (see Connection::untrusted). Use Connection::trusted to remove these limits.

pub fn trusted(self) -> Self

Removes operational limits from this connection. (see Connection::untrusted)

pub fn untrusted(self) -> Self

Sets operational limits for this connection.

Allow INSERT, DELETE only for “queue”, “candidate_subscription”, “subscription”.

Allow UPDATE only for “subscription” user facing settings.

Allow INSERT only for “post”.

Allow read access to all tables.

Allow SELECT, TRANSACTION, SAVEPOINT, and the strftime function.

Deny everything else.

pub fn open_or_create_db(conf: Configuration) -> Result<Self>

Create a database if it doesn’t exist and then open it.

pub fn conf(&self) -> &Configuration

Returns a connection’s configuration.

pub fn load_archives(&self) -> Result<()>

Loads archive databases from Configuration::data_path, if any.

pub fn lists(&self) -> Result<Vec<DbVal<MailingList>>>

Returns a vector of existing mailing lists.

pub fn list(&self, pk: i64) -> Result<Option<DbVal<MailingList>>>

Fetch a mailing list by primary key.

pub fn list_by_id<S: AsRef<str>>( &self, id: S ) -> Result<Option<DbVal<MailingList>>>

Fetch a mailing list by id.

pub fn create_list(&self, new_val: MailingList) -> Result<DbVal<MailingList>>

Create a new list.

pub fn list_posts( &self, list_pk: i64, _date_range: Option<(String, String)> ) -> Result<Vec<DbVal<Post>>>

Fetch all posts of a mailing list.

pub fn list_owners(&self, pk: i64) -> Result<Vec<DbVal<ListOwner>>>

Fetch the owners of a mailing list.

pub fn remove_list_owner(&self, list_pk: i64, owner_pk: i64) -> Result<()>

Remove an owner of a mailing list.

pub fn add_list_owner(&self, list_owner: ListOwner) -> Result<DbVal<ListOwner>>

Add an owner of a mailing list.

pub fn update_list(&mut self, change_set: MailingListChangeset) -> Result<()>

Update a mailing list.

pub fn list_filters( &self, _list: &DbVal<MailingList> ) -> Vec<Box<dyn PostFilter>>

Return the post filters of a mailing list.

impl Connection

pub fn new_smtp_connection( &self ) -> Result<Pin<Box<dyn Future<Output = Result<SmtpConnection>> + Send + 'static>>>

Return an SMTP connection handle if the database connection has one configured.

pub async fn submit( smtp_connection: &mut SmtpConnection, message: &QueueEntry, dry_run: bool ) -> Result<()>

Submit queue items from values to their recipients.