Enum mailpot::mail::ListRequest

pub enum ListRequest {
    Help,
    Subscribe,
    Unsubscribe,
    RetrieveArchive(String, String),
    RetrieveMessages(Vec<String>),
    ChangeSetting(String, bool),
    Other(String),
}

Type of mailing list request.

Variants

Help

Get help about a mailing list and its available interfaces.

Subscribe

Request subscription.

Unsubscribe

Request removal of subscription.

RetrieveArchive(String, String)

Request reception of list posts from a month-year range, inclusive.

RetrieveMessages(Vec<String>)

Request reception of specific mailing list posts from Message-ID values.

ChangeSetting(String, bool)

Request change in subscription settings. See ListSubscription.

Other(String)

Other type of request.