Trait mailpot::mail::message_filters::PostFilter

pub trait PostFilter {
    fn feed<'p, 'list>(
        self: Box<Self>,
        post: &'p mut Post,
        ctx: &'p mut ListContext<'list>
    ) -> Result<(&'p mut Post, &'p mut ListContext<'list>), ()>;
}

Filter that modifies and/or verifies a post candidate. On rejection, return a string describing the error and optionally set post.action to Reject or Defer

Required Methods

fn feed<'p, 'list>(
self: Box<Self>,
post: &'p mut Post,
ctx: &'p mut ListContext<'list>
) -> Result<(&'p mut Post, &'p mut ListContext<'list>), ()>

Feed post into the filter. Perform modifications to post and / or ctx, and return them with Result::Ok unless you want to the processing to stop and return an Result::Err.