Struct mailpot::Connection

pub struct Connection {
    pub connection: DbConnection,
    /* private fields */
}

A connection to a mailpot database.

Fields
connection: DbConnection

The rusqlite connection handle.

Implementations

impl Connection

pub fn insert_to_error_queue(
&self,
env: &Envelope,
raw: &[u8],
reason: String
) -> Result<i64>

Insert a received email into the error queue.

pub fn error_queue(&self) -> Result<Vec<DbVal<Value>>>

Fetch all error queue entries.

pub fn delete_from_error_queue(&mut self, index: Vec<i64>) -> Result<()>

Delete error queue entries.

impl Connection

pub fn insert_post(
&self,
list_pk: i64,
message: &[u8],
env: &Envelope
) -> Result<i64>

Insert a mailing list post into the database.

pub fn post(&self, env: &Envelope, raw: &[u8], _dry_run: bool) -> Result<()>

Process a new mailing list post.

pub fn request(
&self,
list: &DbVal<MailingList>,
request: ListRequest,
env: &Envelope,
_raw: &[u8]
) -> Result<()>

Process a new mailing list request.

pub fn months(&self, list_pk: i64) -> Result<Vec<String>>

Fetch all year and month values for which at least one post exists in yyyy-mm format.

impl Connection

pub fn list_members(&self, pk: i64) -> Result<Vec<DbVal<ListMembership>>>

Fetch all members of a mailing list.

pub fn list_member(&self, list_pk: i64, pk: i64) -> Result<DbVal<ListMembership>>

Fetch mailing list member.

pub fn list_member_by_address(
&self,
list_pk: i64,
address: &str
) -> Result<DbVal<ListMembership>>

Fetch mailing list member by their address.

pub fn add_member(
&self,
list_pk: i64,
new_val: ListMembership
) -> Result<DbVal<ListMembership>>

Add member to mailing list.

pub fn add_candidate_member(
&self,
list_pk: i64,
new_val: ListMembership
) -> Result<i64>

Create membership candidate.

pub fn accept_candidate_member(
&mut self,
pk: i64
) -> Result<DbVal<ListMembership>>

Accept membership candidate.

pub fn remove_membership(&self, list_pk: i64, address: &str) -> Result<()>

Remove a member by their address.

pub fn update_member(
&mut self,
change_set: ListMembershipChangeset
) -> Result<()>

Update a mailing list membership.

impl Connection

pub fn open_db(conf: Configuration) -> Result<Self>

Creates a new database connection.

Connection supports a limited subset of operations by default (see Connection::untrusted). Use Connection::trusted to remove these limits.

pub fn trusted(self) -> Self

Removes operational limits from this connection. (see Connection::untrusted)

pub fn untrusted(self) -> Self

Sets operational limits for this connection.

Allow INSERT, DELETE only for “error_queue”, “queue”, “candidate_membership”, “membership”.

Allow INSERT only for “post”.

Allow read access to all tables.

Allow SELECT, TRANSACTION, SAVEPOINT, and the strftime function.

Deny everything else.

pub fn open_or_create_db(conf: Configuration) -> Result<Self>

Create a database if it doesn’t exist and then open it.

pub fn conf(&self) -> &Configuration

Returns a connection’s configuration.

pub fn load_archives(&self) -> Result<()>

Loads archive databases from Configuration::data_path, if any.

pub fn lists(&self) -> Result<Vec<DbVal<MailingList>>>

Returns a vector of existing mailing lists.

pub fn list(&self, pk: i64) -> Result<DbVal<MailingList>>

Fetch a mailing list by primary key.

pub fn list_by_id<S: AsRef<str>>(
&self,
id: S
) -> Result<Option<DbVal<MailingList>>>

Fetch a mailing list by id.

pub fn create_list(&self, new_val: MailingList) -> Result<DbVal<MailingList>>

Create a new list.

pub fn remove_list_policy(&self, list_pk: i64, policy_pk: i64) -> Result<()>

Remove an existing list policy.

let db = Connection::open_or_create_db(config).unwrap().trusted();
let list_pk = db.create_list(MailingList {
    pk: 0,
    name: "foobar chat".into(),
    id: "foo-chat".into(),
    address: "foo-chat@example.com".into(),
    description: None,
    archive_url: None,
}).unwrap().pk;
db.set_list_policy(
    PostPolicy {
        pk: 0,
        list: list_pk,
        announce_only: false,
        subscriber_only: true,
        approval_needed: false,
        no_subscriptions: false,
        custom: false,
    },
).unwrap();
db.remove_list_policy(1, 1).unwrap();

pub fn remove_list_policy_panic()

let db = Connection::open_or_create_db(config).unwrap().trusted();
db.remove_list_policy(1, 1).unwrap();

pub fn set_list_policy(&self, policy: PostPolicy) -> Result<DbVal<PostPolicy>>

Set the unique post policy for a list.

pub fn list_posts(
&self,
list_pk: i64,
_date_range: Option<(String, String)>
) -> Result<Vec<DbVal<Post>>>

Fetch all posts of a mailing list.

pub fn list_policy(&self, pk: i64) -> Result<Option<DbVal<PostPolicy>>>

Fetch the post policy of a mailing list.

pub fn list_owners(&self, pk: i64) -> Result<Vec<DbVal<ListOwner>>>

Fetch the owners of a mailing list.

pub fn remove_list_owner(&self, list_pk: i64, owner_pk: i64) -> Result<()>

Remove an owner of a mailing list.

pub fn add_list_owner(&self, list_owner: ListOwner) -> Result<DbVal<ListOwner>>

Add an owner of a mailing list.

pub fn update_list(&mut self, change_set: MailingListChangeset) -> Result<()>

Update a mailing list.

pub fn list_filters(
&self,
_list: &DbVal<MailingList>
) -> Vec<Box<dyn PostFilter>>

Return the post filters of a mailing list.