Crate mailpot

Mailing list manager library.

use mailpot::{models::*, Configuration, Connection, SendMail};

let db = Connection::open_or_create_db(config)?.trusted();

// Create a new mailing list
let list_pk = db.create_list(MailingList {
    pk: 0,
    name: "foobar chat".into(),
    id: "foo-chat".into(),
    address: "foo-chat@example.com".into(),
    description: None,
    archive_url: None,
})?.pk;

db.set_list_policy(
    PostPolicy {
        pk: 0,
        list: list_pk,
        announce_only: false,
        subscriber_only: true,
        approval_needed: false,
        no_subscriptions: false,
        custom: false,
    },
)?;

// Drop privileges; we can only process new e-mail and modify memberships from now on.
let db = db.untrusted();

assert_eq!(db.list_members(list_pk)?.len(), 0);
assert_eq!(db.list_posts(list_pk, None)?.len(), 0);

// Process a subscription request e-mail
let subscribe_bytes = b"From: Name <user@example.com>
To: <foo-chat+subscribe@example.com>
Subject: subscribe
Date: Thu, 29 Oct 2020 13:58:16 +0000
Message-ID: <1@example.com>

";
let envelope = melib::Envelope::from_bytes(subscribe_bytes, None)?;
db.post(&envelope, subscribe_bytes, /* dry_run */ false)?;

assert_eq!(db.list_members(list_pk)?.len(), 1);
assert_eq!(db.list_posts(list_pk, None)?.len(), 0);

// Process a post
let post_bytes = b"From: Name <user@example.com>
To: <foo-chat@example.com>
Subject: my first post
Date: Thu, 29 Oct 2020 14:01:09 +0000
Message-ID: <2@example.com>

Hello
";
let envelope =
    melib::Envelope::from_bytes(post_bytes, None).expect("Could not parse message");
db.post(&envelope, post_bytes, /* dry_run */ false)?;

assert_eq!(db.list_members(list_pk)?.len(), 1);
assert_eq!(db.list_posts(list_pk, None)?.len(), 1);

Re-exports

pub extern crate log;

pub extern crate melib;

pub extern crate serde;

pub extern crate serde_json;

Modules

mail

Types for processing new posts: PostFilter, ListContext, MailJob and PostAction.

models

Database models: MailingList, ListOwner, ListMembership, PostPolicy and Post.

Structs

Configuration

The configuration for the mailpot database and the mail server.

Connection

A connection to a mailpot database.

Error

The Error type.

MailtoAddress

A mailto: value.

Enums

ErrorKind

The kind of an error.

SendMail

How to send e-mail.

Traits

ChainedError

This trait is implemented on all the errors generated by the error_chain macro.

Context

Provides the context method for Result.

ResultExt

Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result

Convenient wrapper around std::Result.